kafka/pyproject.toml
[project]
name = "kafka-invoice-streaming"
version = "0.1.0"
description = "Kafka streaming application for invoice data ingestion"
requires-python = ">=3.13"
dependencies = [
    "confluent-kafka>=2.6.0",
    "faker>=33.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
]
